Use `just` instead of `make`
`just` has currently replaced `make` as the command runner for this
project because all but one of the recipes are `.PHONY`; i.e. they don't
refer to actual build artefacts, but to commands to be run. `just` also
provides these notable benefits that align with this project's goals:

* `just` allows a more natural passing of parameters to recipes.
* `just` can be run from any subdirectories in the project (`make` can
  only be run from the root of the project).
* The directory containing the `Justfile` being run can be retrieved
  using `justfile_directory()`, which allows for easier "directory
  agnosticism" in the case of the previous point.
* The string syntax for constructing variables is somewhat more precise
  than `make`.
* The output of the currently-running command is somewhat less noisy
  than `make`, as it squashes consecutive, redundant whitespace.
* Any consistent type of whitespace indentation is allowed in the
  `Justfile`, allowing for an indentation of 4 spaces, which is more
  consistent with the other files in the project.

Note that `make` may be re-introduced into the project in order to
accommodate the use case of building files, if it's needed.